move examples
