+++ folding/src/lib.rs
@@ -11,17 +11,15 @@ use std::{fmt::Debug, hash::Hash};
 // Make available outside the crate to avoid code duplication
 pub use error_term::Side;
 #[cfg(feature = "bn254")]
-pub use example::{Alphas, BaseSponge};
+// pub use example::{Alphas, BaseSponge};
 pub use expressions::{ExpExtension, FoldingCompatibleExpr};
 pub use instance_witness::{Instance, RelaxedInstance, RelaxedWitness, Witness};
 
 pub mod decomposable_folding;
 mod error_term;
-#[allow(dead_code)]
-#[cfg(feature = "bn254")]
-mod example;
 #[cfg(test)]
-mod example_decomposable_folding;
+#[cfg(feature = "bn254")]
+mod examples;
 pub mod expressions;
 mod instance_witness;
 #[cfg(test)]
